requirements.txt
asgiref==3.4.1
Django==3.2.10
django-cors-headers==3.11.0
djangorestframework==3.13.1
Pygments==2.11.2
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
